refactor of app.py and moved code to seperate files for better organization and maintainability
